refactor: extract sort-classes utils that will be useful for `sort-modules`
* Just extracting functions around, nothing really interesting to review
