+++ rules/sort-classes-utils.ts
@@ -9,8 +9,6 @@ import type {
 } from './sort-classes.types'
 import type { CompareOptions } from '../utils/compare'
 
-import { validateNoDuplicatedGroups } from '../utils/validate-groups-configuration'
-import { allModifiers, allSelectors } from './sort-classes.types'
 import { isSortable } from '../utils/is-sortable'
 import { matches } from '../utils/matches'
 
@@ -21,101 +19,6 @@ interface CustomGroupMatchesProps {
   modifiers: Modifier[]
   decorators: string[]
   elementName: string
-}
-/**
- * Cache computed groups by modifiers and selectors for performance
- */
-let cachedGroupsByModifiersAndSelectors = new Map<string, string[]>()
-
-/**
- * Generates an ordered list of groups associated to modifiers and selectors entered
- * The groups are generated by combining all possible combinations of modifiers with one selector in the end
- * Selectors are prioritized over modifiers quantity. This means that
- * `protected abstract override get fields();` should prioritize the 'get-method' group over the 'protected-abstract-override-method' group.
- * @param modifiers List of modifiers associated to the selector, i.e ['abstract', 'protected']
- * @param selectors List of selectors, i.e ['get-method', 'method', 'property']
- */
-export let generateOfficialGroups = (
-  modifiers: Modifier[],
-  selectors: Selector[],
-): string[] => {
-  let modifiersAndSelectorsKey = `${modifiers.join('&')}/${selectors.join('&')}`
-  let cachedValue = cachedGroupsByModifiersAndSelectors.get(
-    modifiersAndSelectorsKey,
-  )
-  if (cachedValue) {
-    return cachedValue
-  }
-  let allModifiersCombinations: string[][] = []
-  for (let i = modifiers.length; i > 0; i--) {
-    allModifiersCombinations = [
-      ...allModifiersCombinations,
-      ...getCombinations(modifiers, i),
-    ]
-  }
-  let allModifiersCombinationPermutations = allModifiersCombinations.flatMap(
-    result => getPermutations(result),
-  )
-  let returnValue: string[] = []
-  for (let selector of selectors) {
-    returnValue = [
-      ...returnValue,
-      ...allModifiersCombinationPermutations.map(
-        modifiersCombinationPermutation =>
-          [...modifiersCombinationPermutation, selector].join('-'),
-      ),
-      selector,
-    ]
-  }
-  cachedGroupsByModifiersAndSelectors.set(modifiersAndSelectorsKey, returnValue)
-  return returnValue
-}
-
-/**
- * Get possible combinations of n elements from an array
- */
-export let getCombinations = (array: string[], number: number): string[][] => {
-  let result: string[][] = []
-
-  let backtrack = (start: number, comb: string[]): void => {
-    if (comb.length === number) {
-      result.push([...comb])
-      return
-    }
-    for (let i = start; i < array.length; i++) {
-      comb.push(array[i])
-      backtrack(i + 1, comb)
-      comb.pop()
-    }
-  }
-
-  backtrack(0, [])
-  return result
-}
-
-/**
- * Get all permutations of an array
- * This allows 'abstract-override-protected-get-method', 'override-protected-abstract-get-method',
- * 'protected-abstract-override-get-method'... to be entered by the user and always match the same group
- * This can theoretically cause performance issues in case users enter too many modifiers at once? 8 modifiers would result
- * in 40320 permutations, 9 in 362880.
- */
-let getPermutations = (elements: string[]): string[][] => {
-  let result: string[][] = []
-  let backtrack = (first: number): void => {
-    if (first === elements.length) {
-      result.push([...elements])
-      return
-    }
-    for (let i = first; i < elements.length; i++) {
-      ;[elements[first], elements[i]] = [elements[i], elements[first]]
-      backtrack(first + 1)
-      ;[elements[first], elements[i]] = [elements[i], elements[first]]
-    }
-  }
-  backtrack(0)
-
-  return result
 }
 
 /**
@@ -255,48 +158,4 @@ export let getCompareOptions = (
     ignoreCase: options.ignoreCase,
     locales: options.locales,
   }
-}
-
-export let validateGroupsConfiguration = (
-  groups: Required<SortClassesOptions[0]>['groups'],
-  customGroups: Required<SortClassesOptions[0]>['customGroups'],
-): void => {
-  let availableCustomGroupNames = new Set(
-    customGroups.map(customGroup => customGroup.groupName),
-  )
-  let invalidGroups = groups
-    .flat()
-    .filter(
-      group =>
-        !isPredefinedGroup(group) && !availableCustomGroupNames.has(group),
-    )
-  if (invalidGroups.length) {
-    throw new Error(`Invalid group(s): ${invalidGroups.join(', ')}`)
-  }
-  validateNoDuplicatedGroups(groups)
-}
-
-let isPredefinedGroup = (input: string): boolean => {
-  if (input === 'unknown') {
-    return true
-  }
-  let singleWordSelector = input.split('-').at(-1)
-  if (!singleWordSelector) {
-    return false
-  }
-  let twoWordsSelector = input.split('-').slice(-2).join('-')
-  let isTwoWordSelectorValid = allSelectors.includes(
-    twoWordsSelector as Selector,
-  )
-  if (
-    !allSelectors.includes(singleWordSelector as Selector) &&
-    !isTwoWordSelectorValid
-  ) {
-    return false
-  }
-  let modifiers = input.split('-').slice(0, isTwoWordSelectorValid ? -2 : -1)
-  return (
-    new Set(modifiers).size === modifiers.length &&
-    modifiers.every(modifier => allModifiers.includes(modifier as Modifier))
-  )
 }
